saved changes

+++ northwind.common/data/Pager.cs
@@ -3,18 +3,18 @@ namespace northwind.common.data
   public class Pager
   {
     private const int DefaultPageSize = 10;
-    public long PageNumber { get; }
+    public int PageNumber { get; }
     public int PageSize { get; }
 
     public long TotalItems { get; }
     
-    public Pager(long pageNumber)
+    public Pager(int pageNumber)
     {
       PageNumber = pageNumber == default ? 1 : pageNumber;
       PageSize = DefaultPageSize;
     }
 
-    public Pager(long pageNumber, long totalItems)
+    public Pager(int pageNumber, int totalItems)
     {
       PageNumber = pageNumber == default ? 1 : pageNumber;
       PageSize = DefaultPageSize;
