+++ northwind.domain/models/Employee.cs
@@ -2,7 +2,7 @@
 {
   public class Employee
   {
-    public long Id { get; set; }
+    public int Id { get; set; }
     public string LastName { get; set; }
     public string FirstName { get; set; }
     public string Title { get; set; }
@@ -18,7 +18,7 @@
     public string Extension { get; set; }
     public byte[] Photo { get; set; }
     public string Notes { get; set; }
-    public long? ReportsTo { get; set; }
+    public int? ReportsTo { get; set; }
     public string PhotoPath { get; set; }
   }
 }
